Refactor: Rename daonode to restapi
