+++ restapi/src/main/java/bisq/restapi/RestApi.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.daonode;
+package bisq.restapi;
 
 
 import bisq.core.account.witness.AccountAgeWitnessService;
