+++ restapi/src/main/java/bisq/restapi/dto/ProofOfBurnDto.java
@@ -15,16 +15,32 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.daonode.error;
+package bisq.restapi.dto;
 
 import lombok.Getter;
+import lombok.ToString;
 
+
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+/**
+ * Minimal data required for Bisq 2 proof of burn use case.
+ * Need to be in sync with the Bisq 2 ProofOfBurnDto class.
+ */
 @Getter
-public class ErrorMessage {
-    private final String error;
+@ToString
+@Schema(title = "ProofOfBurn")
+public class ProofOfBurnDto {
+    private final long amount;
+    private final long time;
+    private final String hash;
+    private final int blockHeight;
 
-    public ErrorMessage(String error) {
-        this.error = error;
+    public ProofOfBurnDto(long amount, long time, String hash, int blockHeight) {
+        this.amount = amount;
+        this.time = time;
+        this.hash = hash;
+        this.blockHeight = blockHeight;
     }
 }
-
