+++ restapi/src/main/java/bisq/restapi/endpoints/AccountAgeApi.java
@@ -15,8 +15,9 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.daonode.endpoints;
+package bisq.restapi.endpoints;
 
+import bisq.core.account.witness.AccountAgeWitness;
 import bisq.core.account.witness.AccountAgeWitnessService;
 
 import java.util.Date;
@@ -27,9 +28,9 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 
 
-import bisq.daonode.RestApi;
-import bisq.daonode.RestApiMain;
-import bisq.daonode.dto.ProofOfBurnDto;
+import bisq.restapi.RestApi;
+import bisq.restapi.RestApiMain;
+import bisq.restapi.dto.ProofOfBurnDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -45,25 +46,26 @@ import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 
 /**
- * Endpoint for getting the signed witness date from a given hash as hex string.
+ * Endpoint for getting the account age date from a given hash as hex string.
  * Used for reputation system in Bisq 2.
- * <a href="http://localhost:8082/api/v1/signed-witness/get-date/dd75a7175c7c83fe9a4729e36b85f5fbc44e29ae">Request with hash</a>
+ * <a href="http://localhost:8082/api/v1/account-age/get-date/dd75a7175c7c83fe9a4729e36b85f5fbc44e29ae">Request with hash</a>
  */
 @Slf4j
-@Path("/signed-witness")
+@Path("/account-age")
 @Produces(MediaType.TEXT_PLAIN)
-@Tag(name = "Signed witness API")
-public class SignedWitnessApi {
-    private static final String DESC_HASH = "The hash of the signed account age witness as hex string";
+@Tag(name = "Account age API")
+public class AccountAgeApi {
+    private static final String DESC_HASH = "The hash of the account age witness as hex string";
+    private final RestApi restApi;
     private final AccountAgeWitnessService accountAgeWitnessService;
 
-    public SignedWitnessApi(@Context Application application) {
-        RestApi restApi = ((RestApiMain) application).getRestApi();
+    public AccountAgeApi(@Context Application application) {
+        restApi = ((RestApiMain) application).getRestApi();
         accountAgeWitnessService = checkNotNull(restApi.getAccountAgeWitnessService());
     }
 
-    @Operation(description = "Request the signed witness date")
-    @ApiResponse(responseCode = "200", description = "The signed witness date",
+    @Operation(description = "Request the account age date")
+    @ApiResponse(responseCode = "200", description = "The account age date",
             content = {@Content(mediaType = MediaType.TEXT_PLAIN,
                     schema = @Schema(allOf = ProofOfBurnDto.class))}
     )
@@ -73,9 +75,9 @@ public class SignedWitnessApi {
                         @PathParam("hash")
                         String hash) {
         long result = accountAgeWitnessService.getWitnessByHashAsHex(hash)
-                .map(accountAgeWitnessService::getWitnessSignDate)
+                .map(AccountAgeWitness::getDate)
                 .orElse(-1L);
-        log.info("SignedWitness sign date for hash {}: {} ({})", hash, result, new Date(result));
+        log.info("Account age for hash {}: {} ({})", hash, result, new Date(result));
         return result;
     }
 }
