+++ restapi/src/main/java/bisq/restapi/endpoints/BondedReputationApi.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.daonode.endpoints;
+package bisq.restapi.endpoints;
 
 import bisq.core.dao.governance.bond.reputation.BondedReputationRepository;
 import bisq.core.dao.state.DaoStateService;
@@ -32,9 +32,9 @@ import lombok.extern.slf4j.Slf4j;
 
 
 
-import bisq.daonode.RestApi;
-import bisq.daonode.RestApiMain;
-import bisq.daonode.dto.BondedReputationDto;
+import bisq.restapi.RestApi;
+import bisq.restapi.RestApiMain;
+import bisq.restapi.dto.BondedReputationDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
